+++ workers/generic-worker/purge_caches_test.go
@@ -273,3 +273,71 @@ func TestPurgeCachesNegativeExitCode(t *testing.T) {
 		t.Fatalf("Was expecting log to contain string %v.", substring)
 	}
 }
+
+// Exit codes specified in OnExitStatus.PurgeCaches should only purge task caches
+func TestPurgeTaskCaches(t *testing.T) {
+	setup(t)
+	mounts := []MountEntry{
+		// requires scope "generic-worker:cache:apple-cache"
+		&WritableDirectoryCache{
+			CacheName: "apple-cache",
+			Directory: filepath.Join("my-task-caches", "apples"),
+		},
+	}
+	payload := GenericWorkerPayload{
+		Command:    returnExitCode(0),
+		MaxRunTime: 30,
+		Mounts:     toMountArray(t, mounts),
+	}
+	defaults.SetDefaults(&payload)
+	td := testTask(t)
+	td.Scopes = []string{"generic-worker:cache:apple-cache"}
+
+	_ = submitAndAssert(t, td, payload, "completed", "completed")
+
+	ensureDirContainsNFiles(t, cachesDir, 1)
+
+	mounts = []MountEntry{
+		// requires scope "generic-worker:cache:banana-cache"
+		&WritableDirectoryCache{
+			CacheName: "banana-cache",
+			Directory: filepath.Join("my-task-caches", "bananas"),
+		},
+	}
+	payload = GenericWorkerPayload{
+		Command:    returnExitCode(123),
+		MaxRunTime: 30,
+		Mounts:     toMountArray(t, mounts),
+		OnExitStatus: ExitCodeHandling{
+			PurgeCaches: []int64{123},
+		},
+	}
+	defaults.SetDefaults(&payload)
+	td = testTask(t)
+	td.Scopes = []string{"generic-worker:cache:banana-cache"}
+
+	_ = submitAndAssert(t, td, payload, "failed", "failed")
+
+	logtext := LogText(t)
+	substring := "[mounts] Removing cache banana-cache from cache table"
+	if !strings.Contains(logtext, substring) {
+		t.Log(logtext)
+		t.Fatalf("Was expecting log to contain string %v.", substring)
+	}
+
+	substring = "[mounts] Deleting cache banana-cache file(s) at"
+	if !strings.Contains(logtext, substring) {
+		t.Log(logtext)
+		t.Fatalf("Was expecting log to contain string %v.", substring)
+	}
+
+	// WritableDirectoryCache should not be unmounted and moved to the cache dir
+	substring = "[mounts] Preserving cache: Moving"
+	if strings.Contains(logtext, substring) {
+		t.Log(logtext)
+		t.Fatalf("Was not expecting log to contain string %v.", substring)
+	}
+
+	// ensure task cache from previous task is still present
+	ensureDirContainsNFiles(t, cachesDir, 1)
+}